add total dollar
